requirements.txt
jinja2>=3.1.2
python-dotenv>=1.0.0
click>=8.1.3
colorama>=0.4.6
